version 0.11 - expanded search field functionality, added background texture, reorganized script.js file
Now, if you search for a marker on a date that it has no event on you will be shown the previous event and the option to jump to it in the timeline will appear. Added paper backround textures and some minor style improvements. Moved all variables to the beginning of script.js.